[reorganization] WIP move single substitution into separate files.\
https://github.com/harfbuzz/harfbuzz/commit/7dfd9e700
